Refactor: DTO 금액 long, 수량 BigDecimal로 전면 전환
- OrderEvent / ExecutionEvent / InboundOrderMessage: price -> long, quantity -> BigDecimal로 변경 및 관련 import/필드 정리.
- OrderStatusEvent: 기존 double 필드 제거하고 price/avgFill/lastFill/limit는 long, remaining/filledQuantity는 BigDecimal로 통일.

+++ matching-engine/src/main/java/com/beyond/MKX/domain/order/entity/OrderEvent.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -22,7 +23,7 @@ public class OrderEvent {
     private String ticker;        // 종목 코드
     private String side;          // 매수(BUY)/매도(SELL)
     private String orderType;     // 주문 유형 (e.g. LIMIT, MARKET, CANCEL)
-    private double price;         // 가격
-    private double quantity;      // 수량
+    private long price;         // 가격
+    private BigDecimal quantity;      // 수량
     private LocalDateTime createdAt; // 생성 시각
 }
